refactor: move useDASAccount to a more appropriate location
