+++ src/components/TransactionItem/TransactionItemCell/index.tsx
@@ -24,7 +24,7 @@ import DecimalCapacity from '../../DecimalCapacity'
 import { parseDiffDate } from '../../../utils/date'
 import Cellbase from '../../Transaction/Cellbase'
 import styles from './index.module.scss'
-import { useDASAccount } from '../../../contexts/providers/dasQuery'
+import { useDASAccount } from '../../../hooks/useDASAccount'
 import { ReactComponent as BitAccountIcon } from '../../../assets/bit_account.svg'
 import { useBoolean, useIsMobile } from '../../../utils/hook'
 import CopyTooltipText from '../../Text/CopyTooltipText'
