+++ src/pages/Address/index.tsx
@@ -13,7 +13,7 @@ import { isAxiosError } from '../../utils/error'
 import { Card, HashCardHeader } from '../../components/Card'
 import { ReactComponent as ShareIcon } from './share.svg'
 import styles from './styles.module.scss'
-import { useDASAccount } from '../../contexts/providers/dasQuery'
+import { useDASAccount } from '../../hooks/useDASAccount'
 
 export const Address = () => {
   const { address } = useParams<{ address: string }>()
