+++ src/pages/Transaction/TransactionCell/index.tsx
@@ -39,7 +39,7 @@ import TransactionReward from '../TransactionReward'
 import Cellbase from '../../../components/Transaction/Cellbase'
 import CopyTooltipText from '../../../components/Text/CopyTooltipText'
 import { useDeprecatedAddr, useIsMobile, useNewAddr } from '../../../utils/hook'
-import { useDASAccount } from '../../../contexts/providers/dasQuery'
+import { useDASAccount } from '../../../hooks/useDASAccount'
 import styles from './styles.module.scss'
 import AddressText from '../../../components/AddressText'
 import { Cell } from '../../../models/Cell'
